backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
asyncpg>=0.29.0
aiosqlite>=0.19.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic-settings>=2.1.0
alembic>=1.13.1
email-validator>=2.1.0.post1
